+++ packages/cryptoplease/lib/app/screens/authenticated/authenticated_flow_screen.dart
@@ -13,7 +13,6 @@ import '../../../core/user_preferences.dart';
 import '../../../features/backup_phrase/module.dart';
 import '../../../features/incoming_split_key_payments/module.dart';
 import '../../../features/outgoing_direct_payments/module.dart';
-import '../../../features/outgoing_direct_payments/presentation/link_listener.dart';
 import '../../../features/outgoing_split_key_payments/module.dart';
 import '../../../features/payment_request/module.dart';
 
@@ -59,9 +58,7 @@ class _AuthenticatedFlowScreenState extends State<AuthenticatedFlowScreen> {
                 const OSKPModule(),
                 const ISKPModule(),
               ],
-              child: ODPLinkListener(
-                child: AutoRouter(key: _homeRouterKey),
-              ),
+              child: AutoRouter(key: _homeRouterKey),
             );
           },
         ),
