+++ packages/cryptoplease/lib/app/screens/authenticated/profile/components/profile_section.dart
@@ -9,7 +9,7 @@ import '../../../../../core/presentation/utils.dart';
 import '../../../../../di.dart';
 import '../../../../../features/incoming_split_key_payments/module.dart';
 import '../../../../../features/legal/flow.dart';
-import '../../../../../features/outgoing_direct_payments/bl/repository.dart';
+import '../../../../../features/outgoing_direct_payments/module.dart';
 import '../../../../../features/outgoing_split_key_payments/bl/repository.dart';
 import '../../../../../features/payment_request/bl/repository.dart';
 import '../../../../../gen/assets.gen.dart';
