+++ packages/cryptoplease/lib/app/screens/authenticated/routes.dart
@@ -3,7 +3,7 @@ import 'package:auto_route/auto_route.dart';
 import '../../../features/app_lock/module.dart';
 import '../../../features/backup_phrase/module.dart';
 import '../../../features/incoming_split_key_payments/module.dart';
-import '../../../features/outgoing_direct_payments/presentation/routes.dart';
+import '../../../features/outgoing_direct_payments/module.dart';
 import '../../../features/outgoing_split_key_payments/presentation/routes.dart';
 import '../../../features/qr_scanner/module.dart';
 import '../../../features/ramp/module.dart';
