+++ packages/cryptoplease/lib/features/outgoing_direct_payments/src/bl/bloc.dart
@@ -7,11 +7,11 @@ import 'package:injectable/injectable.dart';
 import 'package:solana/encoder.dart';
 import 'package:solana/solana.dart';
 
-import '../../../config.dart';
-import '../../../core/amount.dart';
-import '../../../core/tokens/token.dart';
-import '../../../core/transactions/resign_tx.dart';
-import '../../../core/transactions/tx_sender.dart';
+import '../../../../config.dart';
+import '../../../../core/amount.dart';
+import '../../../../core/tokens/token.dart';
+import '../../../../core/transactions/resign_tx.dart';
+import '../../../../core/transactions/tx_sender.dart';
 import 'outgoing_direct_payment.dart';
 import 'repository.dart';
