+++ packages/cryptoplease/lib/features/outgoing_direct_payments/src/bl/outgoing_direct_payment.dart
@@ -2,7 +2,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:solana/encoder.dart';
 import 'package:solana/solana.dart';
 
-import '../../../core/amount.dart';
+import '../../../../core/amount.dart';
 
 part 'outgoing_direct_payment.freezed.dart';
