+++ packages/cryptoplease/lib/features/outgoing_direct_payments/src/presentation/odp_confirmation_screen.dart
@@ -3,18 +3,18 @@ import 'package:decimal/decimal.dart';
 import 'package:dfunc/dfunc.dart';
 import 'package:flutter/material.dart';
 
-import '../../../core/tokens/token.dart';
-import '../../../l10n/device_locale.dart';
-import '../../../l10n/l10n.dart';
-import '../../../ui/amount_keypad/amount_keypad.dart';
-import '../../../ui/app_bar.dart';
-import '../../../ui/bordered_row.dart';
-import '../../../ui/button.dart';
-import '../../../ui/dialogs.dart';
-import '../../../ui/number_formatter.dart';
-import '../../../ui/theme.dart';
-import '../../../ui/usdc_info.dart';
-import 'odp_header.dart';
+import '../../../../core/tokens/token.dart';
+import '../../../../l10n/device_locale.dart';
+import '../../../../l10n/l10n.dart';
+import '../../../../ui/amount_keypad/amount_keypad.dart';
+import '../../../../ui/amount_with_equivalent.dart';
+import '../../../../ui/app_bar.dart';
+import '../../../../ui/bordered_row.dart';
+import '../../../../ui/button.dart';
+import '../../../../ui/dialogs.dart';
+import '../../../../ui/number_formatter.dart';
+import '../../../../ui/theme.dart';
+import '../../../../ui/usdc_info.dart';
 
 class ODPConfirmationScreen extends StatefulWidget {
   const ODPConfirmationScreen({
@@ -87,7 +87,7 @@ class _ScreenState extends State<ODPConfirmationScreen> {
               ),
             ),
             const SizedBox(height: 16),
-            ODPHeader(
+            AmountWithEquivalent(
               inputController: _amountController,
               token: widget.token,
               collapsed: true,
