+++ packages/cryptoplease/test/features/outgoing_direct_payments/bl/bloc_test.dart
@@ -3,9 +3,9 @@ import 'package:cryptoplease/core/amount.dart';
 import 'package:cryptoplease/core/currency.dart';
 import 'package:cryptoplease/core/tokens/token.dart';
 import 'package:cryptoplease/core/transactions/tx_sender.dart';
-import 'package:cryptoplease/features/outgoing_direct_payments/bl/bloc.dart';
-import 'package:cryptoplease/features/outgoing_direct_payments/bl/outgoing_direct_payment.dart';
-import 'package:cryptoplease/features/outgoing_direct_payments/bl/repository.dart';
+import 'package:cryptoplease/features/outgoing_direct_payments/src/bl/bloc.dart';
+import 'package:cryptoplease/features/outgoing_direct_payments/src/bl/outgoing_direct_payment.dart';
+import 'package:cryptoplease/features/outgoing_direct_payments/src/bl/repository.dart';
 import 'package:cryptoplease_api/cryptoplease_api.dart';
 import 'package:dfunc/dfunc.dart';
 import 'package:flutter_test/flutter_test.dart';
